s3-storage-for-3p/backend/utils/security.py
from datetime import datetime, timedelta
from typing import Optional, Dict, Any
from jose import jwt, JWTError
from passlib.context import CryptContext
from fastapi import HTTPException, status, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import boto3
from botocore.exceptions import ClientError
import logging

from config import settings

logger = logging.getLogger(__name__)

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
security = HTTPBearer()


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a password against its hash."""
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    """Generate password hash."""
    return pwd_context.hash(password)


def create_access_token(data: Dict[str, Any], expires_delta: Optional[timedelta] = None) -> str:
    """Create JWT access token."""
    to_encode = data.copy()
    
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.jwt_expire_minutes)
    
    to_encode.update({"exp": expire, "iat": datetime.utcnow()})
    
    encoded_jwt = jwt.encode(
        to_encode,
        settings.jwt_secret_key,
        algorithm=settings.jwt_algorithm
    )
    
    return encoded_jwt


def verify_token(token: str) -> Dict[str, Any]:
    """Verify and decode JWT token."""
    try:
        payload = jwt.decode(
            token,
            settings.jwt_secret_key,
            algorithms=[settings.jwt_algorithm]
        )
        return payload
    except JWTError as e:
        logger.error(f"JWT verification failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Could not validate credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )


def verify_cognito_token(token: str) -> Dict[str, Any]:
    """Verify Cognito JWT token."""
    try:
        client = boto3.client('cognito-idp', region_name=settings.aws_region)
        
        # For production, implement proper JWT verification with Cognito public keys
        # This is a simplified version
        payload = jwt.decode(token, options={"verify_signature": False})
        
        # Verify token is not expired
        if 'exp' in payload:
            exp_timestamp = payload['exp']
            if datetime.utcnow().timestamp() > exp_timestamp:
                raise HTTPException(
                    status_code=status.HTTP_401_UNAUTHORIZED,
                    detail="Token has expired"
                )
        
        return payload
        
    except Exception as e:
        logger.error(f"Cognito token verification failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication token"
        )


async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(security)
) -> Dict[str, Any]:
    """Get current authenticated user from JWT token."""
    token = credentials.credentials
    
    try:
        payload = verify_token(token)
        user_id = payload.get("sub")
        username = payload.get("username")
        
        if user_id is None or username is None:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid authentication credentials"
            )
        
        return {
            "id": user_id,
            "username": username,
            "email": payload.get("email"),
            "cognito_sub": payload.get("cognito_sub")
        }
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Error getting current user: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Could not validate credentials"
        )


def get_user_s3_prefix(user_id: str) -> str:
    """Get S3 prefix for a user based on their Cognito user ID."""
    return f"users/{user_id}/"


def validate_s3_key_access(user_id: str, key: str) -> bool:
    """Validate that a user has access to the specified S3 key."""
    user_prefix = get_user_s3_prefix(user_id)
    
    if not key.startswith(user_prefix):
        logger.warning(f"User {user_id} attempted to access unauthorized key: {key}")
        return False
    
    return True


def sanitize_input(value: str, max_length: int = 255) -> str:
    """Sanitize user input to prevent injection attacks."""
    # Remove null bytes
    value = value.replace('\x00', '')
    
    # Truncate to max length
    if len(value) > max_length:
        value = value[:max_length]
    
    return value.strip()
